requirements.txt
langchain==0.1.20
langchain-google-genai
python-dotenv
